+++ src/main/java/org/example/employee_performance_review_api/application/web/model/request/UpdateEmployeeRequest.java
@@ -11,6 +11,7 @@ import org.example.employee_performance_review_api.domain.model.employee.enums.P
 import org.example.employee_performance_review_api.domain.model.employee.enums.Team;
 import org.example.employee_performance_review_api.infrastructure.web.validation.constraint.AtLeastOneFieldMustBeNotNull;
 
+import java.math.BigDecimal;
 import java.util.List;
 import java.util.UUID;
 
@@ -27,9 +28,9 @@ public class UpdateEmployeeRequest {
   private Team team;
   private PerformanceRating rating;
   private EmployeeType type;
-  private int yearlySalary;
+  private BigDecimal yearlySalary;
   private int employmentTime;
-  private List<Integer> yearlyBonuses;
+  private List<BigDecimal> yearlyBonuses;
 
   public UpdateEmployeeInput toUpdateEmployeeInput(UUID id) {
     return new UpdateEmployeeInput(id, this.firstName, this.lastName, this.team, this.rating, this.type, this.yearlySalary, this.employmentTime, this.yearlyBonuses);
